requirement.txt
PyQt5==5.15.11
olefile==0.47
pillow==12.1.0
numpy==1.26.0
PyYAML==6.0.2
tqdm==4.67.1